backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
ollama
chromadb==0.5.3
sentence-transformers==3.0.1
pypdf==4.2.0
pydantic==2.7.1
python-dotenv==1.0.1
langchain-text-splitters==0.2.2
supabase==2.4.6
httpx==0.27.0
